rename decorator => api_decorator
