+++ st3/lsp_utils/_client_handler/language_handler.py
@@ -1,7 +1,7 @@
 from ..api_wrapper_interface import ApiWrapperInterface
 from ..helpers import log_and_show_message
 from ..server_resource_interface import ServerStatus
-from .decorator import register_decorated_handlers
+from .api_decorator import register_decorated_handlers
 from .interface import ClientHandlerInterface
 from LSP.plugin import ClientConfig
 from LSP.plugin import LanguageHandler
